+++ src/physics/matter-js/typedefs/MatterWorldConfig.js
@@ -26,6 +26,11 @@
  * @property {number} [correction=1] - An optional Number that specifies the time correction factor to apply to the update.
  * @property {function} [getDelta] - This function is called every time the core game loop steps, which is bound to the Request Animation Frame frequency unless otherwise modified.
  * @property {boolean} [autoUpdate=true] - Automatically call Engine.update every time the game steps.
+ * @property {number} [restingThresh=4] - Sets the Resolver resting threshold property.
+ * @property {number} [restingThreshTangent=6] - Sets the Resolver resting threshold tangent property.
+ * @property {number} [positionDampen=0.9] - Sets the Resolver position dampen property.
+ * @property {number} [positionWarming=0.8] - Sets the Resolver position warming property.
+ * @property {number} [frictionNormalMultiplier=5] - Sets the Resolver friction normal multiplier property.
  * @property {(boolean|Phaser.Types.Physics.Matter.MatterDebugConfig)} [debug=false] - Sets if Matter will render to the debug Graphic overlay. As of Phaser 3.22 it should be a `MatterDebugConfig` object instead of a boolean.
  * @property {boolean} [debugShowBody=true] - Should dynamic bodies be drawn to the debug graphic? Deprecated in Phaser 3.22: use a `MatterDebugConfig` object instead.
  * @property {boolean} [debugShowStaticBody=true] - Should static bodies be drawn to the debug graphic? Deprecated in Phaser 3.22: use a `MatterDebugConfig` object instead.